front: rebase and move userSlice in src/reducers/
