+++ front/src/reducers/index.ts
@@ -8,10 +8,7 @@ import { OsrdConfState, OsrdMultiConfState } from 'applications/operationalStudi
 
 import { osrdEditoastApi } from 'common/api/osrdEditoastApi';
 
-import userReducer, {
-  UserState,
-  userInitialState,
-} from 'applications/common/reducer/user/userSlice';
+import userReducer, { UserState, userInitialState } from 'reducers/user';
 import mainReducer, { MainState, MainActions, initialState as mainInitialState } from './main';
 import mapReducer, { MapState, initialState as mapInitialState } from './map';
 import editorReducer, { EditorActions, initialState as editorInitialState } from './editor';
